refactor(frontdesk): split autosync.go by concern
autosync.go held the auto-sync loop alongside the sync-hold handling and
the per-member incomplete/unconfirmed/unreadable bookkeeping. Same
package, pure moves with their doc comments:

- autosync_hold.go: warnIfBuildGateDegraded, holdMemberForSkew,
  closeSyncHold, heldPerLog
- autosync_incomplete.go: incompleteState, divergenceKind and its
  constants, the recordSyncAttempt / markUnconfirmedPush /
  clearUnconfirmedPush / hasUnconfirmedPush / hasBeenPushedSinceReset /
  flagDiverged / markMemberIncomplete / markMemberUnmeasured /
  recordUnreadableHash / clearUnreadableHash / clearMemberIncomplete /
  incompleteSnapshot / shouldSkipIncompleteRetry / resetIncompleteRetries
  methods, unreadableCause, joinCapped, divergenceMessage,
  unmeasuredMessage

autosync.go keeps the constants, the stale tiers, RunAutoSync,
autoSyncOnce, forceAutoSyncNow, primaryConfigHash, convergeFleet, the
section helpers, applyAutoSync and measureMember (1171 -> 667 lines)
and leaves the size allowlist. Tests are unchanged.
