+++ scripts/ci/size-allowlist.txt
@@ -15,7 +15,6 @@ internal/api/failover.go	929
 internal/api/logs.go	809
 internal/api/stats.go	843
 internal/failover/circuitbreaker_test.go	2056
-internal/frontdesk/autosync.go	1171
 internal/frontdesk/autosync_test.go	4254
 internal/model/model_crud_test.go	2399
 internal/provider/discovery_test.go	2235
